Support module CRUD operations in device client
This change adds all the CRUD module operations to the HTTP device
client.

Note that the `Module` model definitions are the same as for the
`iotservice` client and they have been reused instead of redefining them.
Ideally, this model definitions should be in the common package but that
would break backward compatibility.